Start implementing basic scene interface.
