+++ src/effects.py
@@ -1,0 +1,11 @@
+from scenes_base import Effect
+from world import World
+
+
+class ChangeSceneName(Effect):
+
+    def __init__(self, name: str):
+        self._name = name
+
+    def execute(self, world: World) -> None:
+        world.current_scene = self._name
